backend/requirements.txt
openai>=1.30.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
pdfplumber>=0.11.0
reportlab>=4.1.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0